Make some unusual bitrates optional
Some (modern!) systems do not define the
    B76800, B153600, B307200
bitrates, so we have now made them optional. Same goes
for the formerly #ifdef'ed definitions depending on FreeBSD.

This makes the number and value of the bitrates supported
variable, so we now use a variable length list with a terminator.

The C part of eppp_serial should be OK wrt bitrate support.

Potentially the Erlang part of eppp_serial needs some adaptations
as well (to be checked).